Moved `verifyTokenAndAttachUser` into it's own middleware file, shifted Google OAuth request handler to the top of `SessionController`
