+++ apps/augury-backend/src/middlewares/SessionController.ts
@@ -1,29 +1,13 @@
-import { CookieOptions, Request, Response, NextFunction } from 'express';
+import { CookieOptions, Request, Response } from 'express';
 import axios, { AxiosError } from 'axios';
 import qs from 'querystring';
 import { HydratedDocument, Types } from 'mongoose';
-import { signJwt, verifyJwt } from '../config/utils/jwt';
+import { signJwt } from '../config/utils/jwt';
 import ApiError from '../errors/ApiError';
 import User from '../config/interfaces/User';
 import Session from '../config/interfaces/Session';
-import UserModel from '../models/auth/UserModel';
 import SessionModel from '../models/auth/SessionModel';
-import StatusCode from '../config/enums/StatusCode';
-import Severity from '../config/enums/Severity';
-
-const accessCookieOptions: CookieOptions = {
-  maxAge: 900000, // 15 mins
-  httpOnly: true,
-  domain: 'localhost',
-  path: '/',
-  sameSite: 'lax',
-  secure: false,
-};
-
-const refreshCookieOptions: CookieOptions = {
-  ...accessCookieOptions,
-  maxAge: 3.154e10, // 1 year
-};
+import UserModel from '../models/auth/UserModel';
 
 interface GoogleTokensResult {
   access_token: string;
@@ -41,6 +25,65 @@ interface GoogleUserResult {
   given_name: string;
   family_name: string;
   picture: string;
+}
+
+const accessCookieOptions: CookieOptions = {
+  maxAge: 900000, // 15 mins
+  httpOnly: true,
+  domain: 'localhost',
+  path: '/',
+  sameSite: 'lax',
+  secure: false,
+};
+
+const refreshCookieOptions: CookieOptions = {
+  ...accessCookieOptions,
+  maxAge: 3.154e10, // 1 year
+};
+
+export async function googleOauthHandler(req: Request, res: Response) {
+  if (!req?.query?.code || typeof req.query.code !== 'string') {
+    throw new Error('Invalid code provided with OAuth query!');
+  }
+  //get code from query string
+  const code = req.query.code as string;
+  const { id_token, access_token } = await getGoogleOAuthTokens(code);
+  // console.log({ id_token, access_token });
+  //get the id and access token
+
+  //get user with tokens
+  const googleUser = await getGoogleUser(id_token, access_token);
+  // console.log(googleUser);
+  const user = await getUserByGoogleId(googleUser);
+  // console.log(user);
+
+  //create a session
+  const session = await getSession(user._id, user.googleId);
+
+  //create access & refressh token
+  const accessToken = signJwt(
+    { ...user._id, session: session.token },
+    { expiresIn: '15m' } // 15 minutes
+  );
+  const refreshToken = signJwt(
+    { ...user._id, session: session.token },
+    { expiresIn: '1y' } // 1 year
+  );
+  //set cookie
+  res.cookie('accessToken', accessToken, accessCookieOptions);
+  res.cookie('refreshToken', refreshToken, refreshCookieOptions);
+
+  //redirect back to client
+  const clientPort = process.env.CLIENT_PORT || 4200;
+  const url = `${process.env.FRONTEND_URL || 'http://localhost'}:${clientPort}`;
+  // const url = `${
+  //   process.env.FRONTEND_URL || 'http://localhost'
+  // }:${clientPort}/Test.html`;
+  // const serverPort = process.env.SERVER_PORT || 3333;
+  // const url = `${
+  //   process.env.BACKEND_URL || 'http://localhost'
+  // }:${serverPort}/User`;
+  res.redirect(url);
 }
 
 export async function getGoogleOAuthTokens(
@@ -149,86 +192,4 @@ export async function getSession(
   }
 
   return response;
-}
-
-export async function googleOauthHandler(req: Request, res: Response) {
-  if (!req?.query?.code || typeof req.query.code !== 'string') {
-    throw new Error('Invalid code provided with OAuth query!');
-  }
-  //get code from query string
-  const code = req.query.code as string;
-  const { id_token, access_token } = await getGoogleOAuthTokens(code);
-  // console.log({ id_token, access_token });
-  //get the id and access token
-
-  //get user with tokens
-  const googleUser = await getGoogleUser(id_token, access_token);
-  // console.log(googleUser);
-  const user = await getUserByGoogleId(googleUser);
-  // console.log(user);
-
-  //create a session
-  const session = await getSession(user._id, user.googleId);
-
-  //create access & refressh token
-  const accessToken = signJwt(
-    { ...user._id, session: session.token },
-    { expiresIn: '15m' } // 15 minutes
-  );
-  const refreshToken = signJwt(
-    { ...user._id, session: session.token },
-    { expiresIn: '1y' } // 1 year
-  );
-  //set cookie
-  res.cookie('accessToken', accessToken, accessCookieOptions);
-  res.cookie('refreshToken', refreshToken, refreshCookieOptions);
-
-  //redirect back to client
-  const clientPort = process.env.CLIENT_PORT || 4200;
-  const url = `${process.env.FRONTEND_URL || 'http://localhost'}:${clientPort}`;
-  // const url = `${
-  //   process.env.FRONTEND_URL || 'http://localhost'
-  // }:${clientPort}/Test.html`;
-  // const serverPort = process.env.SERVER_PORT || 3333;
-  // const url = `${
-  //   process.env.BACKEND_URL || 'http://localhost'
-  // }:${serverPort}/User`;
-  res.redirect(url);
-}
-
-export async function verifyTokenAndAttachUser(
-  req: Request,
-  _res: Response,
-  next: NextFunction
-) {
-  const accessToken = req?.cookies?.accessToken;
-  if (!accessToken) {
-    throw new ApiError(
-      'No access token found.',
-      StatusCode.UNAUTHORIZED,
-      Severity.LOW
-    );
-  }
-
-  try {
-    // Verify and decode the JWT
-    const verificationResult = verifyJwt(accessToken);
-    if (
-      typeof verificationResult.decoded === 'string' ||
-      !verificationResult?.decoded?.session
-    ) {
-      throw new Error();
-    }
-    const token = verificationResult.decoded.session;
-    const session: Session = await SessionModel.getSessionByToken(token);
-    // Attach the data and pass control to the next middleware/handler
-    req.user = await UserModel.getUser(session.userId);
-    next();
-  } catch (error) {
-    throw new ApiError(
-      'Invalid token structure.',
-      StatusCode.FORBIDDEN,
-      Severity.MED
-    );
-  }
 }
